test(apq): move and update tests
